ylab-breakerbot/breaker-bot/requirements.txt
paho-mqtt==1.6.1
slack-sdk==3.20.2